requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
pyhocon==0.3.61
pycryptodome==3.21.0
numpy==2.1.2
